+++ ajax/billactions.php
@@ -30,6 +30,8 @@
 
 include("../../../inc/includes.php");
 
+Session::checkRight("plugin_order_bill", READ);
+
 header("Content-Type: text/html; charset=UTF-8");
 
 Html::header_nocache();