+++ front/order_item.form.php
@@ -29,6 +29,7 @@
  */
 
 include("../../../inc/includes.php");
+Session::checkLoginUser();
 $item = new PluginOrderOrder_Item();
 
 if (isset($_POST['update'])) {