+++ CHANGELOG.md
@@ -10,12 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Addition of a massive action “Cancel receipt” (in the “attachments” tab)
+- Batch data entry for item generation
 - Add massive selection for delivered items
 - Duplicate the delivery button at the top of the reception list
 
 ### Changed
 
-- Implement `Twig` template for order items list
+- Implement `Twig` template for order items list and associated items
 
 ## [2.10.7] - 2025-03-19